requirements.txt
numpy~=1.23.4
pandas~=1.4.4
yfinance~=0.2.4
dataclasses~=0.8
xlwings~=0.29.1
scipy~=1.10.0
python-dateutil~=2.8.2
PyMuPDF
tqdm~=4.64.1